Created folder for DVDisplayList and all its sub-files
