+++ src/components/DataView/DataViewControlDisplay/DataViewControlDisplayTypes.ts
@@ -1,0 +1,7 @@
+import { DataViewDisplay } from "../DataViewTypes"
+
+export interface DataViewControlDisplayProps {
+	display: string
+	displayOptions: DataViewDisplay[]
+	onDisplayChange(display: string): void
+}
