+++ src/components/DataView/DataViewControlDisplay/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./DataViewControlDisplay";
 export * from "./DataViewControlDisplay";
+export * from "./DataViewControlDisplayTypes";
