+++ src/components/DataView/DataViewDisplays/DataViewDisplays.tsx
@@ -1,9 +1,9 @@
 import GridOnIcon from "@mui/icons-material/GridOn";
 import FormatListBulletedIcon from "@mui/icons-material/FormatListBulleted";
 
-import { DataViewDisplay } from "./DataViewTypes";
-import { default as List } from "./DataViewDisplayList";
-import { default as Grid } from "./DataViewDisplayGrid";
+import { DataViewDisplay } from "../DataViewTypes";
+import { default as List } from "../DataViewDisplayList";
+import { default as Grid } from "../DataViewDisplayGrid";
 
 export const DataViewDisplayList: DataViewDisplay = {
 	name : "list",
