+++ src/components/DataView/DataViewLimit/DataViewControlLimitTypes.ts
@@ -1,0 +1,7 @@
+import { DataViewOnLimitChange } from "../DataViewTypes"
+
+export interface DataViewControlLimitProps {
+	limit: number
+	options: number[]
+	onLimitChange: DataViewOnLimitChange
+}
