+++ src/components/DataView/DataViewLimit/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./DataViewControlLimit";
 export * from "./DataViewControlLimit";
+export * from "./DataViewControlLimitTypes";
